backend/requirements.txt
## STEP 2: BACKEND CONFIGURATION AND DATABASE

### File: backend/requirements.txt

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis
redis==5.0.1
hiredis==2.3.2

# Kafka
aiokafka==0.10.0

# Object Storage (MinIO/S3)
boto3==1.34.28
minio==7.2.3

# LLM & AI
langchain==0.1.4
langchain-community==0.0.16
langgraph==0.0.20
ollama==0.1.6

# Vector Database
weaviate-client==4.4.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
jsonschema==4.20.0

# WebSocket
websockets==12.0

# Monitoring
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
sentry-sdk[fastapi]==1.40.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
httpx==0.26.0
aiofiles==23.2.1
